+++ core/src/main/java/com/klaytn/caver/wallet/Wallet.java
@@ -21,6 +21,7 @@
 package com.klaytn.caver.wallet;
 
 import com.klaytn.caver.crypto.KlayCredentials;
+import com.klaytn.caver.utils.SecureRandomUtils;
 import org.bouncycastle.crypto.digests.SHA256Digest;
 import org.bouncycastle.crypto.generators.PKCS5S2ParametersGenerator;
 import org.bouncycastle.crypto.generators.SCrypt;
@@ -378,7 +379,7 @@ public class Wallet {
         }
     }
 
-    public static byte[] generateRandomBytes(int size) {
+    static byte[] generateRandomBytes(int size) {
         byte[] bytes = new byte[size];
         SecureRandomUtils.secureRandom().nextBytes(bytes);
         return bytes;
